Break out BIOS emulation code into its own file. Remove BOOT_BIN support (who still uses that?). assorted warning fixes.
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@674 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/Core/Src/HW/EXI_DeviceIPL.cpp
@@ -66,7 +66,7 @@ CEXIIPL::CEXIIPL() :
 	if (pStream != NULL)
 	{
 		fseek(pStream, 0, SEEK_END);
-		size_t filesize = ftell(pStream);
+		size_t filesize = (size_t)ftell(pStream);
 		rewind(pStream);
 
 		fread(m_pIPL + 0x001fcf00, 1, filesize, pStream);
@@ -81,7 +81,7 @@ CEXIIPL::CEXIIPL() :
 	if (pStream != NULL)
 	{
 		fseek(pStream, 0, SEEK_END);
-		size_t filesize = ftell(pStream);
+		size_t filesize = (size_t)ftell(pStream);
 		rewind(pStream);
 
 		fread(m_pIPL + 0x001aff00, 1, filesize, pStream);
